Remove Snowflake fairing and associated db module (#17)
This commit also places flaken in the database crate.  A separate type
will wrap up Flaken values so that the non-db stuff doesn't need to know
about it.

Why this is important: It may eventually be useful to change the way we
generate snowflakes.  Flaken is convenient now, but that could change.
The non-database shouldn't have to care about that change; all it needs
is a way to say "give me an ID" and it should get an ID.